third_party/blink/web_tests/virtual/prefetch-reusable/external/wpt/speculation-rules/prefetch/sec-fetch-headers.https-expected.txt

This is a testharness.js-based test.
[FAIL] Sec-Fetch-Dest
  assert_equals: Sec-Fetch-Dest expected "document" but got "empty"
[FAIL] Sec-Fetch-Mode
  assert_equals: Sec-Fetch-Mode expected "navigate" but got "no-cors"
[FAIL] Sec-Fetch-Dest with redirects
  assert_equals: Sec-Fetch-Dest expected "document" but got "empty"
[FAIL] Sec-Fetch-Mode with redirects
  assert_equals: Sec-Fetch-Mode expected "navigate" but got "no-cors"
Harness: the test ran to completion.
